whisper/whisper/requirements.txt
whisper
sounddevice
scipy
openai
python-dotenv
pyperclip